# adapter/grpc/gen/dto/user.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.35.1
// 	protoc        v3.6.1
// source: dto/user.proto

package dto

import (
	resource "github.com/xybor/todennus-backend/adapter/grpc/gen/dto/resource"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	reflect "reflect"
	sync "sync"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type UserValidateRequest struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	Username string `protobuf:"bytes,1,opt,name=username,proto3" json:"username,omitempty"`
	Password string `protobuf:"bytes,2,opt,name=password,proto3" json:"password,omitempty"`
}

func (x *UserValidateRequest) Reset() {
	*x = UserValidateRequest{}
	mi := &file_dto_user_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *UserValidateRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*UserValidateRequest) ProtoMessage() {}

func (x *UserValidateRequest) ProtoReflect() protoreflect.Message {
	mi := &file_dto_user_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use UserValidateRequest.ProtoReflect.Descriptor instead.
func (*UserValidateRequest) Descriptor() ([]byte, []int) {
	return file_dto_user_proto_rawDescGZIP(), []int{0}
}

func (x *UserValidateRequest) GetUsername() string {
	if x != nil {
		return x.Username
	}
	return ""
}

func (x *UserValidateRequest) GetPassword() string {
	if x != nil {
		return x.Password
	}
	return ""
}

type UserValidateResponse struct {
	state         protoimpl.MessageState
	sizeCache     protoimpl.SizeCache
	unknownFields protoimpl.UnknownFields

	User *resource.User `protobuf:"bytes,1,opt,name=user,proto3" json:"user,omitempty"`
}

func (x *UserValidateResponse) Reset() {
	*x = UserValidateResponse{}
	mi := &file_dto_user_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *UserValidateResponse) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*UserValidateResponse) ProtoMessage() {}

func (x *UserValidateResponse) ProtoReflect() protoreflect.Message {
	mi := &file_dto_user_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use UserValidateResponse.ProtoReflect.Descriptor instead.
func (*UserValidateResponse) Descriptor() ([]byte, []int) {
	return file_dto_user_proto_rawDescGZIP(), []int{1}
}

func (x *UserValidateResponse) GetUser() *resource.User {
	if x != nil {
		return x.User
	}
	return nil
}

var File_dto_user_proto protoreflect.FileDescriptor

var file_dto_user_proto_rawDesc = []byte{
	0x0a, 0x0e, 0x64, 0x74, 0x6f, 0x2f, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
	0x12, 0x12, 0x74, 0x6f, 0x64, 0x65, 0x6e, 0x6e, 0x75, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
	0x2e, 0x64, 0x74, 0x6f, 0x1a, 0x17, 0x64, 0x74, 0x6f, 0x2f, 0x72, 0x65, 0x73, 0x6f, 0x75, 0x72,
	0x63, 0x65, 0x2f, 0x75, 0x73, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x4d, 0x0a,
	0x13, 0x55, 0x73, 0x65, 0x72, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71,
	0x75, 0x65, 0x73, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x75, 0x73, 0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65,
	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x75, 0x73, 0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65,
	0x12, 0x1a, 0x0a, 0x08, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x18, 0x02, 0x20, 0x01,
	0x28, 0x09, 0x52, 0x08, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x22, 0x4d, 0x0a, 0x14,
	0x55, 0x73, 0x65, 0x72, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70,
	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x35, 0x0a, 0x04, 0x75, 0x73, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01,
	0x28, 0x0b, 0x32, 0x21, 0x2e, 0x74, 0x6f, 0x64, 0x65, 0x6e, 0x6e, 0x75, 0x73, 0x2e, 0x70, 0x72,
	0x6f, 0x74, 0x6f, 0x2e, 0x64, 0x74, 0x6f, 0x2e, 0x72, 0x65, 0x73, 0x6f, 0x75, 0x72, 0x63, 0x65,
	0x2e, 0x55, 0x73, 0x65, 0x72, 0x52, 0x04, 0x75, 0x73, 0x65, 0x72, 0x42, 0x38, 0x5a, 0x36, 0x67,
	0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x78, 0x79, 0x62, 0x6f, 0x72, 0x2f,
	0x74, 0x6f, 0x64, 0x65, 0x6e, 0x6e, 0x75, 0x73, 0x2d, 0x62, 0x61, 0x63, 0x6b, 0x65, 0x6e, 0x64,
	0x2f, 0x61, 0x64, 0x61, 0x70, 0x74, 0x65, 0x72, 0x2f, 0x67, 0x72, 0x70, 0x63, 0x2f, 0x67, 0x65,
	0x6e, 0x2f, 0x64, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
}

var (
	file_dto_user_proto_rawDescOnce sync.Once
	file_dto_user_proto_rawDescData = file_dto_user_proto_rawDesc
)

func file_dto_user_proto_rawDescGZIP() []byte {
	file_dto_user_proto_rawDescOnce.Do(func() {
		file_dto_user_proto_rawDescData = protoimpl.X.CompressGZIP(file_dto_user_proto_rawDescData)
	})
	return file_dto_user_proto_rawDescData
}

var file_dto_user_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_dto_user_proto_goTypes = []any{
	(*UserValidateRequest)(nil),  // 0: todennus.proto.dto.UserValidateRequest
	(*UserValidateResponse)(nil), // 1: todennus.proto.dto.UserValidateResponse
	(*resource.User)(nil),        // 2: todennus.proto.dto.resource.User
}
var file_dto_user_proto_depIdxs = []int32{
	2, // 0: todennus.proto.dto.UserValidateResponse.user:type_name -> todennus.proto.dto.resource.User
	1, // [1:1] is the sub-list for method output_type
	1, // [1:1] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_dto_user_proto_init() }
func file_dto_user_proto_init() {
	if File_dto_user_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: file_dto_user_proto_rawDesc,
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_dto_user_proto_goTypes,
		DependencyIndexes: file_dto_user_proto_depIdxs,
		MessageInfos:      file_dto_user_proto_msgTypes,
	}.Build()
	File_dto_user_proto = out.File
	file_dto_user_proto_rawDesc = nil
	file_dto_user_proto_goTypes = nil
	file_dto_user_proto_depIdxs = nil
}