# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Todennus - An Open ID Connect and OAuth2 Provider",
        "title": "Todennus API Endpoints",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/users": {
            "post": {
                "description": "Register a new user by providing username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_UserRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Duplicated",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerDuplicatedErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/username/{username}": {
            "get": {
                "description": "Get an user information by user username. \u003cbr\u003e\nRequire scope `read:user.role` to get role of user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get user successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_UserGetByUsernameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerNotFoundErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/validate": {
            "post": {
                "description": "Validate the user credentials and returns the user information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Validate user credentials",
                "parameters": [
                    {
                        "description": "Validation data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validate successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_UserValidateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerInvalidCredentialsErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "description": "Get an user information by user id. \u003cbr\u003e\nRequire scope `read:user.role` to get role of user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get user successfully",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerSuccessResponse-dto_UserGetByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerNotFoundErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.UserGetByIDResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Huy Le Ngoc"
                },
                "id": {
                    "type": "string",
                    "example": "330559330522759168"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.UserGetByUsernameResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Huy Le Ngoc"
                },
                "id": {
                    "type": "string",
                    "example": "330559330522759168"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.UserRegisterRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "s3Cr3tP@ssW0rD"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.UserRegisterResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Huy Le Ngoc"
                },
                "id": {
                    "type": "string",
                    "example": "330559330522759168"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.UserValidateRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "s3Cr3tP@ssW0rD"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.UserValidateResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "Huy Le Ngoc"
                },
                "id": {
                    "type": "string",
                    "example": "330559330522759168"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "standard.SwaggerBadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "error_description": {
                    "type": "string",
                    "example": "invalid field: require string but got int"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "standard.SwaggerDuplicatedErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "duplicated"
                },
                "error_description": {
                    "type": "string",
                    "example": "something has already existed"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "standard.SwaggerInvalidCredentialsErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_credentials"
                },
                "error_description": {
                    "type": "string",
                    "example": "username or password is invalid"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "standard.SwaggerMetadata": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string",
                    "example": "IIJORWQpIvDMzzNf"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-10-23T13:52:29.459752901+07:00"
                }
            }
        },
        "standard.SwaggerNotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "error_description": {
                    "type": "string",
                    "example": "not found user with id 323979471029873"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "standard.SwaggerSuccessResponse-dto_UserGetByIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserGetByIDResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "standard.SwaggerSuccessResponse-dto_UserGetByUsernameResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserGetByUsernameResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "standard.SwaggerSuccessResponse-dto_UserRegisterResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserRegisterResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "standard.SwaggerSuccessResponse-dto_UserValidateResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserValidateResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}